New functionality Fixed: panic when loading map from database

+++ trunk/Server/messages/sendplayerdata.go
@@ -25,7 +25,7 @@ import (
 type SendPlayerData struct {
 	UID			uint64
 	Position	pos.Position
-	Direction	int
+	Direction	uint16
 	Money		int32
 	Name		string
 	Outfit		Outfit
@@ -43,7 +43,7 @@ func (m *SendPlayerData) WritePacket() (*pnet.Packet, os.Error) {
 	packet.AddString(m.Name)
 	packet.AddUint16(uint16(m.Position.X))
 	packet.AddUint16(uint16(m.Position.Y))
-	packet.AddUint16(uint16(m.Direction))
+	packet.AddUint16(m.Direction)
 	packet.AddUint32(uint32(m.Money))
 	
 	packet.AddUint8(uint8(m.Outfit.GetOutfitStyle(OUTFIT_UPPER)))
